% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezcyclone.R
\name{ezcyclone}
\alias{ezcyclone}
\title{Cell cycle phase classification}
\usage{
ezcyclone(sce, organism = "hsa", gene.names = rownames(sce),
  ncores = 1, seed = 100, iter = 1000, min.iter = 100,
  min.pairs = 50, verbose = TRUE)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing expression values, usually counts.}

\item{organism}{hsa or mmu for human or mouse genes.}

\item{gene.names}{Ensembl gene IDs for genes in \code{sce}.}

\item{ncores}{Number of cores.}

\item{seed}{Random seed.}

\item{iter}{An integer scalar specifying the number of iterations for random sampling to obtain a cycle score.}

\item{min.iter}{An integer scalar specifying the minimum number of iterations for score estimation.}

\item{min.pairs}{An integer scalar specifying the minimum number of pairs for cycle estimation.}

\item{verbose}{TRUE/FASLE for specifying whether diagnostics should be printed to screen.}
}
\value{
A list containing phases, scores, and normalized.scores .
}
\description{
Cell cycle phase classification using \pkg{scran} \code{cyclone}
}
