% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_markers.R
\name{find_markers}
\alias{find_markers}
\title{Find marker genes for cell clusters}
\usage{
find_markers(
  sce,
  clusters,
  annot,
  block = NULL,
  design = NULL,
  pval.type = "any",
  lfc = 1,
  test.type = "t",
  direction = "up",
  assay_type = "logcounts",
  ncores = 1,
  fdr_cutoff = 0.25,
  prefix = NULL,
  write = TRUE
)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing expression values, usually counts.}

\item{clusters}{Vector specify cell clusters.}

\item{annot}{gene annotation.}

\item{block}{Vector specify blocking.}

\item{design}{Design matirx.}

\item{pval.type}{A string specifying the type of combined p-value to be computed, i.e., Simes' or IUT.}

\item{lfc}{Log fold-change.}

\item{direction}{Direction of change.}

\item{assay_type}{A string specifying which assay values to use, e.g., "counts" or "logcounts".}

\item{ncores}{Number of cores.}

\item{fdr_cutoff}{FDR cutoff for top marker genes.}

\item{prefix}{Prefix for file name for the QC metrics histograms.}

\item{write}{TRUE/FASLE for whether write the table of filtered cells.}
}
\value{
A data.frame for the statistics and annotation of top marker geness
}
\description{
Find marker genes for cell clusters using \pkg{scran} \code{findMarkers}
}
