% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_genes.r
\name{filter_genes}
\alias{filter_genes}
\title{Filter genes base on avarage counts}
\usage{
filter_genes(sce, cutoff = 0, use_size_factors = FALSE,
  prefix = NULL, plot = TRUE, write = TRUE)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing expression values, usually counts.}

\item{cutoff}{The cutoff of avarage counts for filtering.}

\item{use_size_factors}{a logical scalar specifying whetherthe size factors in \code{object} should be used to construct effective library sizes.}

\item{prefix}{Prefix for file name for the QC metrics histograms.}

\item{plot}{TRUE/FASLE for whether plot the QC metrics histograms.}

\item{write}{TRUE/FASLE for whether write the table of filtered cells.}
}
\value{
A SingleCellExperiment object.
}
\description{
Filter genes that have avarage counts fewer than \code{cutoff)}
}
