% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_genes.r
\name{filter_genes}
\alias{filter_genes}
\title{Filter genes base on avarage counts}
\usage{
filter_genes(
  sce,
  cutoff = 0,
  ncores = 1,
  prefix = NULL,
  plot = TRUE,
  write = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing expression values, usually counts.}

\item{cutoff}{The cutoff of avarage counts for filtering. \code{cutoff} should be >= 0.}

\item{ncores}{Number of cores.}

\item{prefix}{Prefix for file name for the QC metrics histograms.}

\item{plot}{TRUE/FALSE for whether to plot the QC metrics histograms.}

\item{write}{TRUE/FALSE for whether to write the table of filtered cells.}

\item{verbose}{TRUE/FALSE for specifying if diagnostics should be printed to screen.}
}
\value{
A SingleCellExperiment object.
}
\description{
Filter genes that have avarage counts fewer than \code{cutoff)}
}
