% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_markers.R
\name{find_markers}
\alias{find_markers}
\title{Find marker genes for cell clusters}
\usage{
find_markers(
  sce,
  clusters,
  annot,
  block = NULL,
  lfc = 1,
  test.type = c("t", "wilcox", "binom"),
  direction = c("up", "down"),
  pval.type = c("any", "all"),
  assay_type = c("logcounts", "counts"),
  ncores = 1,
  fdr_cutoff = 0.25,
  prefix = NULL,
  write = TRUE
)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing expression values, usually counts.}

\item{clusters}{Vector specify cell clusters.}

\item{annot}{gene annotation.}

\item{block}{Vector specify blocking.}

\item{lfc}{Log fold-change.}

\item{test.type}{String specifying the type of pairwise test to perform -
a t-test with \code{"t"}, a Wilcoxon rank sum test with \code{"wilcox"}, 
or a binomial test with \code{"binom"}.}

\item{direction}{Direction of change.}

\item{pval.type}{A string specifying how p-values are to be combined across pairwise comparisons for a given group/cluster.}

\item{assay_type}{A string specifying which assay values to use, e.g., "counts" or "logcounts".}

\item{ncores}{Number of cores.}

\item{fdr_cutoff}{FDR cutoff for top marker genes.}

\item{prefix}{Prefix for file name for the QC metrics histograms.}

\item{write}{TRUE/FALSE for whether to write the table of filtered cells.}
}
\value{
A data.frame for the statistics and annotation of top marker geness
}
\description{
Find marker genes for cell clusters using \pkg{scran} \code{findMarkers}
}
