% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pairs.R
\name{find_pairs}
\alias{find_pairs}
\title{Cell cycle training step of the pair-based prediction}
\usage{
find_pairs(sce, G1, S, G2M, genes = NULL)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing expression values, usually counts.}

\item{G1}{vector of column indices of G1 genes.}

\item{S}{vector of column indices of S genes.}

\item{G2M}{vector of column indices of G2M genes.}

\item{genes}{Ensembl gene IDs for genes in \code{sce}.}
}
\value{
A list containing training gene pairs for G1, S, and G2M phase.
}
\description{
Cell cycle pairs creation using \pkg{scran} \code{sandbag}.
}
\details{
Genes in all three phases are required.
}
