% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pairs.R
\name{find_pairs}
\alias{find_pairs}
\title{Cell cycle training step of the pair-based prediction}
\usage{
find_pairs(sce, genes.list = genes.list, G1 = G1, S = S, G2M = G2M)
}
\arguments{
\item{genes.list}{Ensembl gene IDs for genes in \code{sce}.}

\item{G1}{vector of column number having G1 genes expression}

\item{S}{vector of column number having S genes expression}

\item{G2M}{vector of column number having G2M genes expression}
}
\value{
A list containing training gene pairs for G1, S and G2M phase.
}
\description{
Cell cycle pairs creation using \pkg{scran} \code{sandbag}
}
