% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pairs.R
\name{find_pairs}
\alias{find_pairs}
\title{Cell cycle training step of the pair-based prediction}
\usage{
find_pairs(sce, G1, S, G2M, genes = NULL)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing expression values, usually counts.}

\item{G1}{Vector of column indices of G1 genes.}

\item{S}{Vector of column indices of S genes.}

\item{G2M}{Vector of column indices of G2M genes.}

\item{genes}{Ensembl gene IDs for genes in \code{sce}. If \code{genes} are set to \code{NULL} , genes will be taken
from \code{sce} .}
}
\value{
A list containing training gene pairs for \code{G1}, \code{S}, and \code{G2M} phase.
}
\description{
Cell cycle pairs creation using \pkg{scran} \code{sandbag}.
}
\details{
Genes in all three phases are required.

Minimum 1000 genes should overlap between \code{sce} and \code{genes}.

Function throws error if no pairs are found.
}
