% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_metrics.r
\name{qc_metrics}
\alias{qc_metrics}
\title{Quality control on the cells}
\usage{
qc_metrics(
  sce,
  sym_col = "symbol",
  by_nmads = TRUE,
  thresholds = c(3, 3, 3),
  ncores = 1,
  prefix = NULL,
  plot = TRUE,
  write = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing expression values, usually counts.}

\item{sym_col}{The column name for the gene symbols in \code{rowData(sce)}.}

\item{by_nmads}{TRUE/FALSE; should number of median absolute deviation be used as thresholds?}

\item{thresholds}{Numeric vector of length 3. If \code{by_nmads} is \code{FALSE}, the elements represent thresholds for
total number of reads, number of detected genes, percentage of mitochondrial genes. If \code{by_nmads} is \code{TRUE},
thresholds for all 3 elements is calculated in terms of number of median absolute deviations, which should not exceed 5.}

\item{ncores}{Number of cores.}

\item{prefix}{Prefix for file name for the QC metrics histograms.}

\item{plot}{TRUE/FALSE for whether to plot the QC metrics histograms.}

\item{write}{TRUE/FALSE for whether to write the table of filtered cells.}

\item{verbose}{TRUE/FALSE for specifying if diagnostics should be printed to screen.}
}
\value{
A SingleCellExperiment object.
}
\description{
Quality control on the cells i.e. filter cells by library sizes, number of expressed genes and mitochondrial gene
proportion, visualize the QC metrics by histograms.
}
