% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tech_trend.R
\name{tech_trend}
\alias{tech_trend}
\title{Make a technical trend & decompose the gene-level variance}
\usage{
tech_trend(sce, dispersion = 0, span = 0.4, block = NA,
  design = NA, assay_type = "logcounts", ncores = 1, prefix = NULL,
  plot = TRUE)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing expression values, usually counts.}

\item{dispersion}{A numeric scalar specifying the dispersion for the NB distribution.
If zero, a Poisson distribution is used.}

\item{span}{the parameter \eqn{\alpha} which controls the degree of
    smoothing.}

\item{block}{A factor specifying the blocking level for each cell.}

\item{design}{A numeric matrix describing the uninteresting factors contributing to expression in each cell.
Alternatively, a single factor can be supplied for one-way layouts.}

\item{ncores}{Number of cores.}

\item{prefix}{Prefix for file name for the QC metrics histograms.}

\item{plot}{TRUE/FASLE for whether plot the QC metrics histograms.}
}
\value{
A function accepting a mean log-expression as input and returning the variance of the log-expression as the output
}
\description{
Make a technical trend using \pkg{scran} \code{makeTechTrend} and decompose the gene-level variance using \pkg{scran} \code{decomposeVar}.
}
