% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tech_trend.R
\name{tech_trend}
\alias{tech_trend}
\title{Make a technical trend & decompose the gene-level variance}
\usage{
tech_trend(
  sce,
  dispersion = 0,
  assay_type = "logcounts",
  ncores = 1,
  size.factors = 1,
  prefix = NULL,
  plot = TRUE
)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing expression values, usually counts.}

\item{dispersion}{A numeric scalar specifying the dispersion for the NB distribution.
If zero, a Poisson distribution is used.}

\item{assay_type}{A string specifying which assay values to use, e.g., "counts" or "logcounts".}

\item{ncores}{Number of cores.}

\item{size.factors}{A numeric vector of size factors.}

\item{prefix}{Prefix for file name for the QC metrics histograms.}

\item{plot}{TRUE/FALSE for whether to plot the QC metrics histograms.}
}
\value{
A function accepting a mean log-expression as input and returning the variance of the log-expression as the output.
}
\description{
Make a technical trend using \pkg{scran} \code{makeTechTrend} and decompose the gene-level variance using
\pkg{scran} \code{modelGeneVar}.
}
